+++ stream-access-control/access-control-core/src/main/java/com/backbase/stream/configuration/AccessControlConfiguration.java
@@ -84,7 +84,8 @@ public class AccessControlConfiguration {
     @Bean
     public AccessGroupService accessGroupService(
         com.backbase.dbs.accesscontrol.api.service.ApiClient accessControlApiClient,
-        com.backbase.dbs.user.api.service.ApiClient usersApiClient) {
+        com.backbase.dbs.user.api.service.ApiClient usersApiClient,
+        BackbaseStreamConfigurationProperties configurationProperties) {
 
         UserManagementApi usersApi = new UserManagementApi(usersApiClient);
         UserQueryApi userQueryApi = new UserQueryApi(accessControlApiClient);
@@ -97,7 +98,8 @@ public class AccessControlConfiguration {
         ServiceAgreementApi serviceAgreementApi = new ServiceAgreementApi(accessControlApiClient);
         ServiceAgreementsApi serviceAgreementsApi = new ServiceAgreementsApi(accessControlApiClient);
         return new AccessGroupService(usersApi, userQueryApi, accessControlUsersApi, dataGroupApi, dataGroupsApi,
-            functionGroupApi, functionGroupsApi, serviceAgreementQueryApi, serviceAgreementApi, serviceAgreementsApi);
+            functionGroupApi, functionGroupsApi, serviceAgreementQueryApi, serviceAgreementApi, serviceAgreementsApi,
+            configurationProperties);
     }
 
     @Bean