+++ stream-legal-entity/legal-entity-core/src/test/java/com/backbase/stream/LegalEntitySagaTest.java
@@ -1,11 +1,15 @@
 package com.backbase.stream;
 
+import static com.backbase.stream.service.UserService.REMOVED_PREFIX;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import com.backbase.dbs.user.api.service.v2.model.GetUser;
+import com.backbase.dbs.user.api.service.v2.model.GetUsersList;
 import com.backbase.dbs.user.api.service.v2.model.Realm;
 import com.backbase.stream.configuration.LegalEntitySagaConfigurationProperties;
 import com.backbase.stream.legalentity.model.BaseProductGroup;
@@ -31,9 +35,11 @@ import com.backbase.stream.service.AccessGroupService;
 import com.backbase.stream.service.LegalEntityService;
 import com.backbase.stream.service.UserProfileService;
 import com.backbase.stream.service.UserService;
+import java.util.ArrayList;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CopyOnWriteArrayList;
@@ -425,6 +431,103 @@ class LegalEntitySagaTest {
             Arrays.asList("100000-Default PG", "100000-Mixed PG"),
             productGroupTaskProcessingOrder
         );
+    }
+
+    @Test
+    void deleteLegalEntity_usersPrefixedWithRemovedNotProcessed() {
+        String leExternalId = "someLeExternalId";
+        String leInternalId = "someLeInternalId";
+        String customSaExId = "someCustomSaExId";
+
+        LegalEntity legalEntity = new LegalEntity().internalId(leInternalId).externalId(leExternalId)
+            .parentExternalId(leExternalId);
+        ServiceAgreement sa = new ServiceAgreement().externalId(customSaExId).creatorLegalEntity(leExternalId);
+
+        when(legalEntityService.getMasterServiceAgreementForExternalLegalEntityId(leExternalId)).thenReturn(
+            Mono.just(sa));
+        when(legalEntityService.getLegalEntityByExternalId(leExternalId)).thenReturn(Mono.just(legalEntity));
+
+        Long totalUsers = 5L;
+
+        List<GetUser> users = getUsers(totalUsers.intValue());
+        users.get(0).setExternalId(REMOVED_PREFIX + "user0");
+        users.get(1).setExternalId(REMOVED_PREFIX + "user1");
+        GetUsersList getUsersList1 = new GetUsersList();
+        getUsersList1.setTotalElements(totalUsers);
+        getUsersList1.users(users);
+
+        when(userService.getUsersByLegalEntity(eq(leInternalId), anyInt(), anyInt()))
+            .thenReturn(Mono.just(getUsersList1));
+
+        when(accessGroupService.removePermissionsForUser(any(), any())).thenReturn(Mono.empty());
+
+        when(accessGroupService.deleteFunctionGroupsForServiceAgreement(any())).thenReturn(Mono.empty());
+        when(accessGroupService.deleteAdmins(any())).thenReturn(Mono.empty());
+        when(userService.archiveUsers(any(), any())).thenReturn(Mono.empty());
+        when(legalEntityService.deleteLegalEntity(any())).thenReturn(Mono.empty());
+
+        Mono<Void> result = legalEntitySaga.deleteLegalEntity(leExternalId);
+        result.block();
+
+        verify(accessGroupService, times(3)).removePermissionsForUser(any(), any());
+    }
+
+    @Test
+    void deleteLegalEntity_processesPaginatedListOfUsers() {
+        String leExternalId = "someLeExternalId";
+        String leInternalId = "someLeInternalId";
+        String customSaExId = "someCustomSaExId";
+
+        LegalEntity legalEntity = new LegalEntity().internalId(leInternalId).externalId(leExternalId)
+            .parentExternalId(leExternalId);
+        ServiceAgreement sa = new ServiceAgreement().externalId(customSaExId).creatorLegalEntity(leExternalId);
+
+        when(legalEntityService.getMasterServiceAgreementForExternalLegalEntityId(leExternalId)).thenReturn(
+            Mono.just(sa));
+        when(legalEntityService.getLegalEntityByExternalId(leExternalId)).thenReturn(Mono.just(legalEntity));
+
+        Long totalUsers = 22L;
+
+        List<GetUser> users = getUsers(totalUsers.intValue());
+        GetUsersList getUsersList1 = new GetUsersList();
+        getUsersList1.setTotalElements(totalUsers);
+        getUsersList1.users(users.subList(0, 10));
+
+        GetUsersList getUsersList2 = new GetUsersList();
+        getUsersList2.setTotalElements(totalUsers);
+        getUsersList2.users(users.subList(10, 20));
+
+        GetUsersList getUsersList3 = new GetUsersList();
+        getUsersList3.setTotalElements(totalUsers);
+        getUsersList3.users(users.subList(20, 22));
+
+        when(userService.getUsersByLegalEntity(eq(leInternalId), anyInt(), anyInt()))
+            .thenReturn(Mono.just(getUsersList1))
+            .thenReturn(Mono.just(getUsersList2))
+            .thenReturn(Mono.just(getUsersList3));
+
+        when(accessGroupService.removePermissionsForUser(any(), any())).thenReturn(Mono.empty());
+
+        when(accessGroupService.deleteFunctionGroupsForServiceAgreement(any())).thenReturn(Mono.empty());
+        when(accessGroupService.deleteAdmins(any())).thenReturn(Mono.empty());
+        when(userService.archiveUsers(any(), any())).thenReturn(Mono.empty());
+        when(legalEntityService.deleteLegalEntity(any())).thenReturn(Mono.empty());
+
+        Mono<Void> result = legalEntitySaga.deleteLegalEntity(leExternalId);
+        result.block();
+
+        verify(userService, times(3)).getUsersByLegalEntity(eq(leInternalId), anyInt(), anyInt());
+    }
+
+    private List<GetUser> getUsers(int amount) {
+        List<GetUser> users = new ArrayList<>();
+
+        for (var i = 0; i < amount; i++) {
+            GetUser user = new GetUser();
+            user.externalId("user_" + i);
+            users.add(user);
+        }
+        return users;
     }
 
     @NotNull