+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+
+## [2.69.0]
+### Added
+- Ability to configure what Function Group type needs to be deleted. This can be configured through `backbase.stream.deletion.functionGroupItemType`. With values `NONE` (default) or `TEMPLATE`.
+
+### Fixed
+- When deleting a legal entity, it now will iterate over all found users to be deleted.
+
 ## [2.68.0]
 ### Added
 - Added debitCards to Savings Account in LegalEntity spec.
@@ -10,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [2.67.0]
 ### Added
 - Added cardDetails to Credit Card in LegalEntity spec.
+
 
 ## [2.66.0]
 ### Changed